- removed old semi-integration tests
